mission-control-ui/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
redis==4.1.4
requests==2.28.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
gunicorn==20.1.0
python-dotenv==0.21.1